requirements.txt
tinydb
openai
python-dotenv
python-telegram-bot
flask
gunicorn
